Removed outer Assignment folder as it caused issues last time
