--- Assignment/Assignment/Assignment.csproj
@@ -1,7 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-	<PropertyGroup>
-		<LangVersion>12.0</LangVersion>
-	</PropertyGroup>
-
-</Project>
